Introduce a new block backend called 'ntnx'
This ntnx backend provides a 100% userspace datapath for tapdisk to access
storage on Nutanix hyperconverged appliances. It uses libfrodoiscsi.so, a
highly efficient library for iSCSI protocol handling and SCSI command
passthrough. The .so is shipped separately by Nutanix and is not part of
the tapdisk distribution. Therefore, tapdisk does not link with the .so
at build time and instead the ntnx backend dlopen()'s the .so at runtime.

block-ntnx.c, which connects tapdisk with libfrodoiscsi.so, is contributed
by Nutanix under the same 3-clause BSD license as rest of tapdisk.