[CA-220467] CID-11258 Unchecked return value from library
CID-11258 is complaining that we do not check the result of send
after calling it. Because of this, I have added the check for the
return value, not just to see if there is an error, but also to see
if the send did not manage to send all the data and it needs to be
called again.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>