CA-215114: Resource leak in vhd_util_coalesce_open_output()
There is a case where vhd_open() succeeds (0 is returned) but
the 'src' and 'dst' header block sizes are different.
'vhd_util_coalesce_open_output()' returns an error, but 'dst' is
not properly closed.

These 2 distinct cases (opening the vhd and checking if it is
correct) are now handled seperately.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>
Reviewed-by: Germano Percossi <germano.percossi@citrix.com>