[CA-220467] CID-62178 CID-57555 Read or Write to pointer after free
The problem is that tapdisk_nbdserver_reqs_init was calling
tapdisk_nbdserver_free_request which was potentially freeing
client without reporting any error, leading to a possible
segmentation fault.

The fix is to extract a new function from
tapdisk_nbdserver_free_request which only does the assignment of the req
without freeing the client and only call this new function
from tapdisk_nbdserver_reqs_init since we know we will never need to
free the client from that init context.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>