CA-222244: Fix resource leaks in vhd_shift_metadata()
There is a case where 'buf' has memory allocated, the pointer
is not copied to locators[i] yet and we fail. 'buf' will be
leaked.

Fix:
- Initialize 'buf' to 'NULL'
- 'posix_memalign()' will set 'buf' to 'NULL' in case of error;
  remove redundant 'NULL'
- Set 'buf' to 'NULL' at the end of every iteration in the loop
- Free 'buf' in 'out'

In the second for loop, if 'vhd_check_for_clobber()' fails, all
allocated memory is leaked.

Fix:
Instead of returning -EINVAL, set 'err' to it and go to 'out'.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>
Reviewed-by: Stefano Panella <stefano.panella@citrix.com>