Add a check to force O_DSYNC if needed
In order to be able to open the VHD leaf with O_DSYNC
we now have a check for the existence of a file.

The file is /etc/tapdisk_use_dsync

If the file exists the option O_DSYNC is added to O_DIRECT.

This is not a public interface, but just an hardcoded debug
interface.

Signed-off-by: Stefano Panella <stefano.panella@citrix.com>