CA-225067: tap_ctl_{spawn,create}: move tapdisk to cgroup slice
The '-c' option allows the user to specify which cgroup slice the new tapdisk
process should be put in. If unspecified, it is moved to the default cgroup
slice.

This avoids the need to rely on the cgrulesengd daemon which fails to do its job
if the time it takes to write to the tasks file exceeds the lifespan of the
parent tapdisk (before it forks and exits). On a 4.4 kernel, it can take several
milliseconds to write to this file, so that kind of failure is very likely.

Failure to move the tapdisk process to the desired cgroup is logged but ignored.

The compile-time option -DTAP_CTL_NO_DEFAULT_CGROUP_SLICE can disable the
default behaviour of moving spawned tapdisk processes to the system's default
cgroup.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>