Propagate correctly const attribute from constant strings
Constant strings are allocate on .rodata section so accessing
as write may cause a crash. In order to make this harder propagate
the const attribute so to cause compiler errors.
Some API (like getopt) for historic reasons accepts a char** type
instead of a const char** so a cast is required there.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>