XOP-944 Fix crashes and errors in stats
The low-level routine which fills the stats report buffer has a couple
of bugs.

First, the check for whether the existing buffer size was exceeded which
inspects the return code for vsnprintf() has an off-by-one error. The
symptom of this will be that if the string to be added to the buffer
would exactly match the end of the current buffer (not including the
terminating NULL), then it would be truncated and the last character
would be replaced by a NULL. Then the next call which appended to the
buffer would extend it as normal, leaving a NULL in the middle of the
buffer, which would seriously confuse any JSON parsers trying to decode
its contents.

Secondly, in the event that the buffer needs to be extended, the
vnsprintf() can be called twice. Since it is varadic and not defended
with va_copy()/va_end(), then in the event of a buffer needing extension
this would result in a segfault or random uninitialised memory being
used to populate the stats buffer.

Thirdly, one should not use void pointers for pointer arithmetic (gcc
permits this but it is not standard C).

Fix the above bugs and populate the skeleton test framework with some
tests to detect the first two cases.

Signed-off-by: Tim Smith <tim.smith@citrix.com>