CA-227096 Allow stashed passed fds to overwrite with the same name
When a filedescriptor is passed with a name, the name and fd are stored
in a free slot and later looked up by name. This means that if ever a
new fd is passed for a name, it will be invisible as the retriever will
return fd stored against the first matching name, which will be the
first instance.

To fix this, allow passed fds to overwrite the fd stored against an
existing name with the new fd.

Also, the current time was stored against each entry, but this is never
used so remove it. And correct the spelling of
tdnbd_retrieve_passed_fd()

Signed-off-by: Tim Smith <tim.smith@citrix.com>