CA-299175: Deal with events on non-existent paths
In rare cases tapback is asked to process a watch where the path in
xenstore has been deleted before the first notification. In this case
the logic could get confused and run on to call frontend() with a NULL
argument, causing frontend() to assert().

tapback_backend_probe_device() is documented to either create or remove
a device. If it finds itself doing neither, simply warn and return an
error.