CA-299175 Guard against NULL devices
We can be sent a notification for a device which has been removed. We
need to guard against this to prevent assertions, but cannot insist that
at least one of (create or remove) is true because there is additional
state-change behaviour.

Update the comments and add a check for device to prevent assertions.