CA-304562 Stats use-after-free
It is possible for attempts to be made to access through
blkif->stats.xenvbd after the memory it points to has been unmapped by
calling tapdisk_xenblkif_stats_destroy().

NULL the pointer as the memory is unmapped and check it before each
use. If tapdisk_xenblkif_stats_destroy() has been called, we shouldn't
be interested in those updates any more anyway.