Add major version to LIBBLOCKCRYPTO_NAME
Adding the major version helps packaging in OpenXT.  OpenEmbedded, used
by OpenXT, packages libblockcrypto.so in the -dev package - only
libblockcrypto.so.0 as a symlink to libblockcrypto.so.0.0.0 are in the
regular package.  This leads to a runtime failure when dlopen can't find
the library - "failed to load crypto support library"

It may not matter here, but it's more defensive to specify the major
version.  That way a major version change will result in an error
instead of being used when there can be incompatible changes.

libblockcrypto.so should already be a symlink to libblockcrypto.so.0.0.0
anyway, so there should be no functional change.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>