CA-327558: add user-space io_getevents() implementation
From spdk.io, which is BSD licensed.
FIO/QEMU use a similar approach.

Instead of making a syscall to check for completion events,
just read the page where the kernel has put the completion information.

This only support io_getevents where min_nr==0, which is what we used.
i.e. the userspace io_getevents doesn't support blocking until an event arrives,
but we use an eventfd for that already.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>