CA-327558: iocb union helper functions
We only use IO_CMD_PREAD and IO_CMD_PWRITE so far,
and in that case nbytes is always in bytes.
However IO_CMD_{PREADV/PWRITEV} is not in bytes: it is the number
of iovec entries in the iocb, and have to sum all of these up to
get the result in bytes.

Add a helper function that does that and use it everywhere,
instead of accessing the union without checking the tag.

The only function accessing the union directly should be iocb-optimize.h

Signed-off-by: Edwin Török <edvin.torok@citrix.com>