feat(vhd/canonpath): add support for DRBD resource path
Preserve symlinks when "/dev/drbd/by-res/<RES_NAME>/<VOLUME>" paths are used.
It's interesting when a parent is set to a VHD file to avoid usage of
"/dev/drbdXXXX" paths where XXXX is the device minor number.

We can have:
> vhd-util query -n /dev/drbd/by-res/volume-d803817e-6113-4b89-abdf-d4ab39d63e8c/0 -p
/dev/drbd/by-res/volume-d5f5a079-0529-4b97-95be-9a859056d523/0

instead of:
> vhd-util query -n /dev/drbd/by-res/volume-d803817e-6113-4b89-abdf-d4ab39d63e8c/0 -p
/dev/drbd/./drbd1004

If a DRBD resource exists on many hosts with a resource name equal to an UUID, it can be used
independently of the DRBD device minor number on each host.

Signed-off-by: Ronan Abhamon <ronan.abhamon@vates.fr>