nbdserver: Read options in NBD_OPT_GO
nbd-client is sending NBD_OPT_GO, and tapdisk promptly finds an invalid
magic value of 0x00000000.  tapdisk is not reading the option sent with
NBD_OPT_GO, so it gets read by the subsequent message handler.  Drain
the option when handling NBD_OPT_GO.

Also the status messages was for NBD_OPT_GO.  Fix that.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>