nbdserver: Handle newstyle negotiation from client object
For many of the un-implemented NBD options, err was not set, so the
client would not get a response and just hang.  They need to set ret to
return an error.

If a newstyle negotiation fails, the client is not shutdown as the FD
remains open.  newstyle is actually handled in a callback on the same FD
as the client.  This makes it a little racy, as two events are using the
same FD.  There is a single NBD server object, but potentially multiple
clients.  The use of single handshake_fd means things could break if two
clients connect at the same time.

Re-work to allocate the client and use client_fd for all the callbacks.
The client is first used with tapdisk_nbdserver_handshake_cb.  When that
is triggered, it is unregistered.  This streamlines the error exiting,
and the callback can't run more than once.  If negotiation completes
successfully, tapdisk_nbdserver_enable_client() registers the new
callback.

The changes to using a client spread through to multiple locations.

With this change, handshake_fd can be removed.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>