td-blkif: Fix misleading indentation
GCC warns:
td-blkif.c: In function ‘tapdisk_xenblkif_connect’:
td-blkif.c:595:5: error: this ‘if’ clause does not guard... [-Werror=misleading-indentation]
  595 |     if (unlikely(err))
      |     ^~
td-blkif.c:598:9: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘if’
  598 |         td_blkif->stoppolling_event = tapdisk_server_register_event(
      |         ^~~~~~~~

This file is mostly spaces, but tabs are the preferred style.  Convert
spaces to tabs in the area, which resolved the warning though making the
file less consistent.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>