td-req: Fix misleading indentation
GCC warns:
td-req.c: In function ‘tapdisk_xenblkif_complete_request’:
td-req.c:547:5: error: this ‘if’ clause does not guard... [-Werror=misleading-indentation]
  547 |     if (final)
      |     ^~
td-req.c:550:9: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘if’
  550 |         if (unlikely(processing_barrier_message))
      |         ^~

Convert all the identation from spaces to tabs for all of
tapdisk_xenblkif_complete_request().  This is more than strictly
necessary, but it makes the whole function uniform.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>