block-vhd: Always load crypto for -DOPEN_XT
OpenXT doesn't pass in the encryption key, but loads it later when
opening the VHD itself.  Therefore it needs to always load the crypto
library to have it available.  Re-arrange to make that the case.

While doing this, swap the if branches, to let the condition be a
positive check.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>