CA-366614: clean up old tapdisk log files
Whilst tapdisk primarily logs via syslog and into daemon.log it also
creates logs in /var/log/blktap, which are rotated imperfectly due to
the filename having the process id in it. This can result in a build
up of very old log files taking significant space in the log partition
which will not be freed by the standard logrotate mechanism. Remove
any files not modified in the last month.

Signed-off-by: Mark Syms <mark.syms@citrix.com>