Silence GCC stringop-truncation warnings
The stringop-truncation warning will fire when there is an
attempt to strncpy N characters into a destination string of
length N, including the null character. The safe_strncpy()
function does this for us by copying N characters and truncating
at length N-1, but the previous strncpy() call is technically
"incorrect" as GCC sees it. Reduce the sz by one to resolve
these warnings. This behavior should be functionally equivalent
to the original.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>