CP-37221: Only read data where present in bitmap
Reading an entire 2MB allocation block from the node being coalesced
when only a small portion of it is required incurs a significant
performance overhead and makes it hard for the coalesce process to
complete if there is any significant amount of IO occuring in the
guest. Update vhd_util_coaleasce_block to read only those extents
marked as present in the bitmap when the BATMAP is not stuffed.

Signed-off-by: Mark Syms <mark.syms@citrix.com>