CP-46763: When VBD queue is not ready call the driver completion
When the tapdisk-vbd layer receives a request to forward an IO request
it checks if the VBD queue is ready. If it not ready, either being
quiesced or less likely is already quiesced, it fails the request with
EBUSY to allow the entire request to be retired later. When doing so
it directly calls its own internal completion function and does not
notify the forwarding driver of the status which might result in that
driver still believing the request to be in progress when it is asked
to close.

Signed-off-by: Mark Syms <mark.syms@citrix.com>