tapback: Move backend to InitWait earlier
This is needed for libxl integration.

tapdisk waits to transition to InitWait until after "hotplug-status" is
connected.  However libxl doesn't run its hotplug scripts, which write
"hotplug-status", until the backend transitions to InitWait.  With both
sides waiting, progress is not made and connecting the blktap device
times out and fails.

Make tapback transition to InitWait earlier to resolve this issue under
libxl.  Place the transition to InitWait in
tapback_backend_create_device() after the xenstore feature nodes have
been written.  InitWait is a signal to the frontend that such nodes have
been written.  This matches blkback's behaviour.  It should also be fine
since tapback still won't advance to Connected without the other setup
like physical-device-path and hotplug-status.

VBDs can be reconnected.  When pv-grub is used, it connects the VBD,
loads the kernel, disconnects the VBD.  It then re-sets the frontend
state to XenbusStateInitialising so that the new kernel will find and
connect the VBD.

tapback and blkback handle this case differently.  When blkback observes
the frontend transition to XenbusStateInitialising, and the backend is
XenbusStateClosed, the backend transitions to XenbusStateInitWait.

When tapback observes the frontend transition to
XenbusStateInitialising, the backend checks for hotplug_status_connected
to be true before switching XenbusStateInitWait.  For tapback, this
serves a second purpose for setting XenbusStateInitWait initially as
well.

With tapback setting XenbusStateInitWait earlier, follow the blkback
model to transition to XenbusStateInitWait when the backend was
previously in XenbusStateClosed.  The end result shouldn't change, but
it will remove an extra write and watch trigger for XenbusStateInitWait.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>