CA-384162: Send completions for all partial responses
After tapdisk-vbd forwards a td_request_t to the next driver layer
that driver is at liberty to decompose the request into smaller
sub-requests, e.g. to accomodate spanning a VHD allocation block
boundary. If this occurs, the individual sub-requests will complete
independently and call the request callback in the higher driver
layer.

The callback handler in block-valve, `__valve_complete_treq`
does not correctly handle the condition if an intermediate request
completes with an error but the final request which decrements the
outstanding sector count to 0 suceeds. This results in a completion
being sent to the next layer up the stack reporting that the entire IO
request is complete when actually it is not and thus at some point
later the guest VM will detect corruption to its disk structure,
filesystem etc.

Correct this behaviour by forwarding the completion event to the next
layer up the stack.

Signed-off-by: Mark Syms <mark.syms@citrix.com>