CA-390490: allocate vhd request before locking bat
In the possible event that there were no available vhd_request
available it was possible that the operation would be failed having
marked the BAT as locked with no attempt to unwind that lock. As it
happens unwinding the lock could potentially be difficult as it is
permitted for multiple requests which are updating the same newly
created allocation block to be submitted so, just because the
VHD_FLAG_REQ_UPDATE_BAT flag is set it is not certain that this
request was the first to lock. To simplify this, first acquire the
vhd_request which has simple cleanup if it fails, then acquire the BAT
lock and release the vhd_request if this fails.

Signed-off-by: Mark Syms <mark.syms@citrix.com>