Once loaded, free the crypto_interface only on exit
The crypto interface structure holds essentially global state for the
VHD encryption support, it should only be freed on shutdown and not
when VHDs themseleves are closed. Closing, and freeing,
crypto_interface when closing VHDs means that when there is more than
one VHD open in the process the xts_tfm for all except the first VHD
closed will be leaked.

Signed-off-by: Mark Syms <mark.syms@citrix.com>