CA-392151 lcache.c uses wrong buffer size definition
To compute its buffer size, block-lcache.c was using
MAX_SEGMENTS_PER_REQ, which has the same value as
BLKIF_MAX_SEGMENTS_PER_REQ (11 at present).

However, td-req.c used BLKIF_MMAX_SEGMENTS_PER_REQ (Note the additional
"M") which has a different value (32). Thus it is possible for a td-req
entity passed to the "tapdisk_lcache" driver to be bigger than the
driver can handle. But the driver doesn't check this and assumes it has
enough space for the memcpy(), resulting in... trouble.

Correct the calculation in block-lcache.c to match the calculation in
td-req.c by sharing a new TD_REQ_BUFFER_SIZE macro via td_req.h

To help prevent further confusion, remove the local definition of
MAX_SEGMENTS_PER_REQ and use BLKIF_MAX_SEGMENTS_PER_REQ in place of it,
so that things which are supposed to be the same at least have the same
name. Rename BLKIF_MMAX_SEGMENTS_PER_REQ to
BLKIF_MAX_BUFFER_SEGMENTS_PER_REQUEST to make it more clear that this is
supposed to be different.

Signed-off-by: Tim Smith <tim.smith@cloud.com>