CA-401174 Filedescriptor leak in NBD option negotiation
If the initial NBD server identification is sent, the NBD server
registers a callback for events on the filedescriptor and awaits further
activity. When this occurs after an NBD_CLOSE, as we allow a connection
to be re-used for a new session, we must make sure to close the
filedescriptor if the session is NOT re-used.

A previous fix plugged an FD leak which occurred when the NBDMAGIC
failed to be sent fully at session initialisation. However, there is
another leak which can occur when the NBDMAGIC is send, but not read,
and the client closes the connection during the callback-driven phase of
option negotiation.

Close that FD here whenever option negotiation fails.

Also, when there is an error during client communication which results
in a client shutdown, the FD needs to be closed there too.

Signed-off-by: Tim Smith <tim.smith@cloud.com>