CA-403297: when storing read-through do not mirror
When the lcache driver has read from the remote storage and returned
the data to the caller it then stores the data into the local vhdcache
by forwarding a write to the vhdcache file. As the vhdcache file is
opened in TD_VBD_SECONDARY_MIRROR mode this also results in the data
being unnecessarily written back to the remote where it has just been
read from.

Signed-off-by: Mark Syms <mark.syms@cloud.com>