CA-387574: Fix build time coverage gathering
Directing the profile-dir to an out of tree location results in gcc 12
creating paths with the separators of the absolute source path
converted to # characters. When subsequently using gcov or lcov to
analyse the coverage this produces errors and no coverage data is
produced. As the collect-test-results.sh pulls these files back into
the build tree and published them from there it isn't necessary to
write them elsewhere and this avoids the decorated paths being created.

Signed-off-by: Mark Syms <mark.syms@cloud.com>