CA-414626: initialize cpumon even if lowmem initialization failed
Low memory monitoring and CPU usage monitoring are used for different
purposes: if we fail to initialize one, we should still try to
initialize the other.

Low memory monitoring is used to reduce queue depth / buffers allocated
when we have a lot of disks attached on the same host.

CPU usage monitoring is used to determine when it is safe to enter
busy polling mode: when Dom0 is already under high load then we don't enter
busy polling mode to allow other processes to use the CPU.
If CPU usage monitoring is not initialized then tapdisk never enter
polling mode, which reduces guest I/O performance.

Low memory monitoring currently fails to initialize with a newer kernel CA-400404,
which exposed this latent performance bug in tapdisk.

After this change tapdisk now correctly enters polling mode.
There is a separate bug that it won't stay in polling mode long enough,
but that'll be fixed in a separate PR.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>