CA-416464: return BLKIF_RSP_EOPNOTSUPP for EOPNOTSUPP
Returning BLKIF_RSP_ERROR if an operation is not supported does not
allow the frontend to exercise any discretion on how to handle the
response and may lead to an operating system crash. As different
backends may support different feature sets and we might, during a
migration, switch backends, an in-flight request might be issued (or
reissued) which is then not supported by this backend.

Signed-off-by: Mark Syms <mark.syms@cloud.com>